requirements.txt
# -----------------------------------------------------------------------------
# written by: Lawrence McDaniel
#             https://lawrencemcdaniel.com
#
# usage:      Python requirements for AWS Lambda functions. Create a virtual
#             environment in the root of this repository named `venv`. Terraform
#             modules will look for and include these requirements in the zip
#             packages for each Python-based Lambda function.
# -----------------------------------------------------------------------------

# dev and test
# ------------
boto3==1.33.6
pytest==7.4.3
pytest_mock==3.12.0

# Code linters, formatters, and security scanners
# ------------
black==23.11.0
flake8==6.1.0
flake8-coding==1.3.2
pre-commit==3.6.0
isort==5.12.0
mypy==1.7.1
pylint==3.0.2
bandit==1.7.6
pydocstringformatter==0.7.3
tox==4.11.4
codespell==2.2.6

# production
# ------------
python-dotenv==1.0.0
openai==1.4.0
langchain-experimental==0.0.43
pinecone-client==2.2.4
